requirements.txt
fastapi
uvicorn
transformers
torch
torchvision
pillow
scikit-learn
requests
